transfer logic to component
