+++ src/containers/translate-scripts/index.tsx
@@ -1,8 +1,27 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 
 import Script from 'next/script';
 
+import { printModeState } from 'store/print-mode';
+
+import { useRecoilValue } from 'recoil';
+
 const TranslateScripts = () => {
+  const isPrintingMode = useRecoilValue(printModeState);
+
+  useEffect(() => {
+    if (typeof window !== 'undefined') {
+      if (isPrintingMode) {
+        const langSelector = document.querySelector('#tx-live-lang-container');
+        langSelector?.classList.add('hidden-langselector');
+      }
+      if (!isPrintingMode) {
+        const langSelector = document.querySelector('#tx-live-lang-container');
+        langSelector?.classList.remove('hidden-langselector');
+      }
+    }
+  }, [isPrintingMode]);
+
   return (
     <>
       <Script
